+++ src/DOC_GPU/DOCGPU_Kernels.h
@@ -178,13 +178,16 @@ bool generateRandomIntArrayDevice(cudaStream_t stream,
 bool whatDataIsInCentroid(cudaStream_t stream,
 						  unsigned int dimBlock,
 						  bool* output,
-						  unsigned long* count,
+						  unsigned int* count,
 						  float* data,
-						  unsigned long* centroids,
+						  unsigned int* centroids,
 						  bool* dimensions,
 						  const float width,
-						  const unsigned long point_dim,
-						  const unsigned long no_data_p);
+						  const unsigned int point_dim,
+						  const unsigned int no_data_p);
+
+
+
 
 
 #endif
